Add tasky as a task management plugin
All functions are grouped under the new Todo group (<leader>x). There
todos can be added and managed. Toggling the display is under the view
options group (even though the current todo is not yet shown). The
keymap to check a todo as done uses folke/snacks to show a custom
message with the next todo.