Update Lualine
The left side now shows the mode and a small icon indicating the current
state of Tabnine. This is more compact than the old state label.
Then the filename is shown (without Git diffs). And then the Git
repository name, the current branch and the diff. The branch name is
extracted with a custom function. The custom function for the branch
name truncates it after the ticket id, should it start with a Jira
ticket id.

The right side shows the 30 minute timer (not every timer) should it be
running and the current tasky todo item should there be one (only if the
label is not turned off in the view options).
The second slot is empty and the last slot shows the current position in
the file.